+++ NEWS.md
@@ -1,5 +1,7 @@
 # nanoscopeAFM 2.0.5
 
+* allow `trimPeaks`, which removes extreme points for better contrast
+* fix graphing of channels > 1
 * add line profiles
 
 # nanoscopeAFM 2.0.4